+++ src/app/modules/shared/models/enum/data-object-building-mode.model.ts
@@ -1,4 +1,0 @@
-export class DataObjectBuildingMode {
-  static INCLUDE_CONTEXT = 'ic';
-  static EXCLUDE_CONTEXT = 'ec';
-}